feat: added info about materials into the spheres

+++ src/gpu_rtx/rgb.rs
@@ -3,7 +3,7 @@ use std::fmt;
 
 pub type RGB = Vector<3>;
 
-fn linear_to_gamma(linear_component: f64) -> f64 {
+fn linear_to_gamma(linear_component: f32) -> f32 {
     if linear_component > 0.0 {
         return linear_component.sqrt();
     }
